pyproject.toml
[tool.poetry]
name = "cdk-workshop"
version = "0.1.0"
description = "CDK Workshop"
authors = ["Oscar Orellana"]

[tool.poetry.dependencies]
python = "3.10.*"
constructs = "10.*"
aws-cdk-lib = "*"
boto3 = "*"
pandas = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
pylint = "*"
flake8 = "*"
black = "*"
isort = "*"
coverage = "*"
coverage-badge = "*"

[tool.black]
extend-exclude = "lambda/layers"

[tool.flake8]
max-line-length = 160
exclude = [".git","__pycache__","build","dist","venv","cdk.out",".pytest_cache"]

[tool.isort]
profile="black"
multi_line_output=8
src_paths=["cdk_workshop/**","lambda/**"]
skip=["venv"]